test(talk): split reusable gateway consult coverage
